refactor: remove onTag slot, call builder directory during tag/snap
